Refactor src into multiple files
