+++ constants/game.h
@@ -1,0 +1,20 @@
+//------------------------------------------------------------------------------
+// Copyright (C) Taylor Woll. All rights reserved.
+// Licensed under the MIT license. See LICENSE.txt file in the project root for
+// full license information.
+//------------------------------------------------------------------------------
+
+#ifndef __JEZNES_CONSTANTS_GAME_H__
+#define __JEZNES_CONSTANTS_GAME_H__
+
+#define MAX_PLAYERS 2
+#define PLAYER_SPEED 0x2
+
+#define MAX_BALLS 20
+#define BALL_SPEED 0x1
+#define BALL_WIDTH 8
+#define BALL_HEIGHT 8
+
+#define TARGET_CLEARED_TILE_PERCENTAGE 75
+
+#endif  // __JEZNES_CONSTANTS_GAME_H__
