+++ constants/sprites.h
@@ -1,0 +1,17 @@
+//------------------------------------------------------------------------------
+// Copyright (C) Taylor Woll. All rights reserved.
+// Licensed under the MIT license. See LICENSE.txt file in the project root for
+// full license information.
+//------------------------------------------------------------------------------
+
+#ifndef __JEZNES_CONSTANTS_SPRITES_H__
+#define __JEZNES_CONSTANTS_SPRITES_H__
+
+#define SPRITE_INDEX_PLAYFIELD_LINE_HORIZ_BASE 0x80
+#define SPRITE_INDEX_PLAYFIELD_LINE_VERT_BASE 0x90
+
+#define SPRITE_INDEX_CURSOR 0x17
+
+#define SPRITE_INDEX_PAUSE_BASE 0x23
+
+#endif  // __JEZNES_CONSTANTS_SPRITES_H__
