+++ screens/game_over.h
@@ -1,0 +1,39 @@
+//------------------------------------------------------------------------------
+// Copyright (C) Taylor Woll. All rights reserved.
+// Licensed under the MIT license. See LICENSE.txt file in the project root for
+// full license information.
+//------------------------------------------------------------------------------
+
+#ifndef __JEZNES_SCREENS_GAME_OVER_H__
+#define __JEZNES_SCREENS_GAME_OVER_H__
+
+// Cursor sprite location for the game over screen.
+#define GAME_OVER_CURSOR_RETRY_X 0x70
+#define GAME_OVER_CURSOR_RETRY_Y 0xc0
+#define GAME_OVER_CURSOR_QUIT_X 0x70
+#define GAME_OVER_CURSOR_QUIT_Y 0xd0
+
+// Game over screen RLE graphics data.
+const unsigned char game_over_screen[] = {
+    0x01,0x70,0x01,0x62,0x41,0x43,0x01,0x17,0x42,0x70,0x01,0x05,0x40,0x70,0x01,0x17,
+    0x50,0x70,0x01,0x05,0x40,0x70,0x0a,0x04,0x10,0x08,0x70,0x12,0x19,0x08,0x15,0x70,
+    0x01,0x0d,0x50,0x70,0x01,0x05,0x40,0x70,0x01,0x17,0x50,0x70,0x01,0x05,0x61,0x53,
+    0x01,0x17,0x62,0x70,0x01,0x05,0x40,0x70,0x01,0x17,0x50,0x70,0x01,0x05,0x40,0x70,
+    0x0f,0x08,0x19,0x08,0x0f,0x70,0x1e,0x1e,0x70,0x01,0x0e,0x50,0x70,0x01,0x05,0x40,
+    0x70,0x01,0x17,0x50,0x70,0x01,0x05,0x40,0x70,0x16,0x06,0x12,0x15,0x08,0x70,0x1e,
+    0x01,0x05,0x70,0x01,0x0a,0x50,0x70,0x01,0x05,0x40,0x70,0x01,0x17,0x50,0x70,0x01,
+    0x05,0x61,0x53,0x01,0x17,0x62,0x70,0x01,0x05,0x40,0x00,0x01,0x17,0x50,0x70,0x01,
+    0x05,0x40,0x00,0x01,0x17,0x50,0x70,0x01,0x05,0x40,0x00,0x01,0x17,0x50,0x70,0x01,
+    0x05,0x40,0x00,0x01,0x17,0x50,0x70,0x01,0x05,0x40,0x00,0x01,0x17,0x50,0x70,0x01,
+    0x05,0x40,0x00,0x01,0x17,0x50,0x70,0x01,0x05,0x40,0x00,0x01,0x17,0x50,0x70,0x01,
+    0x05,0x40,0x00,0x01,0x17,0x50,0x70,0x01,0x05,0x61,0x53,0x01,0x17,0x62,0x70,0x01,
+    0x05,0x40,0x70,0x01,0x17,0x50,0x70,0x01,0x05,0x40,0x70,0x01,0x0b,0x15,0x08,0x17,
+    0x15,0x1c,0x70,0x01,0x06,0x50,0x70,0x01,0x05,0x40,0x70,0x01,0x17,0x50,0x70,0x01,
+    0x05,0x40,0x70,0x01,0x0b,0x14,0x18,0x0c,0x17,0x70,0x01,0x07,0x50,0x70,0x01,0x05,
+    0x40,0x70,0x01,0x17,0x50,0x70,0x01,0x05,0x51,0x53,0x01,0x17,0x52,0x70,0x01,0x22,
+    0xea,0xfa,0x01,0x05,0xba,0xee,0xf0,0xf0,0xfc,0xff,0x01,0x02,0xbb,0xee,0x00,0x00,
+    0x0c,0x0f,0x01,0x02,0xbb,0xee,0x5f,0x01,0x05,0xbb,0xee,0x55,0x01,0x05,0xbb,0xee,
+    0xf5,0x01,0x05,0xbb,0xee,0x00,0x01,0x05,0xbb,0x0e,0x0f,0x01,0x05,0x0b,0x01,0x00
+};
+
+#endif  // __JEZNES_SCREENS_GAME_OVER_H__
