+++ screens/pause.h
@@ -1,0 +1,14 @@
+//------------------------------------------------------------------------------
+// Copyright (C) Taylor Woll. All rights reserved.
+// Licensed under the MIT license. See LICENSE.txt file in the project root for
+// full license information.
+//------------------------------------------------------------------------------
+
+#ifndef __JEZNES_SCREENS_PAUSE_H__
+#define __JEZNES_SCREENS_PAUSE_H__
+
+// PAUSE letter starting location for the pause screen.
+#define PAUSE_LETTER_BASE_X 0x68
+#define PAUSE_LETTER_BASE_Y 0x48
+
+#endif  // __JEZNES_SCREENS_PAUSE_H__
